pyproject.toml
[project]
name = "project-charon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=6.30.0",
    "pytz>=2025.2",
    "strands-agents-tools>=0.2.1",
    "strands-agents[litellm]>=1.0.1",
    "tools>=1.0.2",
    "tree>=0.2.4",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.4",
    "ty>=0.0.1a15",
]
